split the brokers hierarchy into manageable chunks
